Updates kubernetes deployment config and containers for v1.14 and v1.15
Deployment was verified using a single node minikube setup running both kubernetes versions 1.14 and 1.15 and on a 3 node EKS cluster in AWS (running 1.14). The new docker containers referenced in the config were also built using the latest version of the repository. This commit also fixes a minor misspelling in the Crawler.scala file.